passaggio a float

+++ src/org/generation/italy/BigliettoDelTreno.java
@@ -19,15 +19,15 @@ public class BigliettoDelTreno {
 		    System.out.println("Età: " + eta);
 		    System.out.println("kilometri scelti: " + kilometri);
 
-		    float costoBiglietto = kilometri * 0.21;
+		    float costoBiglietto = kilometri * 0.21f;
 		    System.out.println("costo del biglietto: " + costoBiglietto);
 		    
 		    if(eta < 18) {
-		    	double costoBigliettoGiovani = costoBiglietto - costoBiglietto * 20 / 100;
+		    	float costoBigliettoGiovani = costoBiglietto - costoBiglietto * 20 / 100;
 		    	System.out.println("prezzo con lo sconto young: " + costoBigliettoGiovani);
 		      
 		    }else if (eta > 65) {
-		    	double costoBigliettoAnziani = costoBiglietto - costoBiglietto * 40 / 100;
+		    	float costoBigliettoAnziani = costoBiglietto - costoBiglietto * 40 / 100;
 		    	System.out.println("prezzo con lo sconto young: " + costoBigliettoAnziani);
 		    }
